requirements.txt
absl-py==0.15.0
aiohttp==3.8.0
aiosignal==1.2.0
ansi2html==1.6.0
anyio==3.3.2
argon2-cffi==21.1.0
astunparse==1.6.3
async-timeout==4.0.0
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
Brotli==1.0.9
cachetools==4.2.4
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.3
cycler==0.10.0
dash==2.0.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.5.0
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
facenet-pytorch==2.5.2
Flask==2.0.2
Flask-Compress==1.10.1
flatbuffers==2.0
fonttools==4.28.2
frozenlist==1.2.0
fsspec==2021.11.0
future==0.18.2
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.1
h5py==3.5.0
idna==3.2
imageio==2.13.2
imbalanced-learn==0.8.1
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.2
joblib==1.1.0
json5==0.9.6
jsonschema==4.0.1
keras==2.7.0
Keras-Preprocessing==1.1.2
keras-resnet==0.2.0
kiwisolver==1.3.2
libclang==12.0.0
lxml==4.7.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.5.0
matplotlib-inline==0.1.3
mistune==0.8.4
mtcnn==0.1.1
multidict==5.2.0
multipledispatch==0.6.0
multitasking==0.0.10
nbclassic==0.3.2
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
nltk==3.6.7
notebook==6.4.4
numpy==1.21.2
oauthlib==3.1.1
opencv-python==4.5.4.60
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.3
pandocfilters==1.5.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
plotly==5.5.0
portpicker==1.4.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.19.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyDeprecate==0.3.1
pygame==2.1.0
Pygments==2.10.0
PyOpenGL==3.1.5
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyrr==0.10.3
pyrsistent==0.18.0
python-dateutil==2.8.2
pytorch-ignite==0.4.8
pytorch-lightning==1.5.0
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==22.3.0
regex==2022.1.18
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
resnet==0.1
retrying==1.3.3
rsa==4.7.2
scikit-image==0.19.0
scikit-learn==1.0
scipy==1.7.1
seaborn==0.11.2
Send2Trash==1.8.0
setuptools-scm==6.3.2
six==1.16.0
sklearn==0.0
sniffio==1.2.0
soupsieve==2.3.1
tenacity==8.0.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.21.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tifffile==2021.11.2
tomli==1.2.2
toolz==0.11.1
torch==1.7.1+cu110
torchaudio==0.7.2
torchmetrics==0.6.0
torchvision==0.8.2+cu110
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
typing-extensions==3.10.0.2
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.0.2
widgetsnbextension==3.6.0
wikipedia==1.4.0
wrapt==1.13.3
yarl==1.7.2
